% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_and_reporting_functions.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize Gene Regulatory Network with Pie Charts}
\usage{
visualize_network(
  g,
  female_res,
  male_res,
  vertex.size = 5,
  vertex.label.cex = 0.8,
  ...
)
}
\arguments{
\item{g}{An igraph network object.}

\item{female_res}{Differential expression results for females.}

\item{male_res}{Differential expression results for males.}

\item{vertex.size}{Size of the network nodes.}

\item{vertex.label.cex}{Text size for vertex labels.}

\item{...}{Additional graphical parameters.}
}
\value{
The modified igraph object with visualization attributes.
}
\description{
Plots a network with nodes represented by pie charts that display male and female effects.
}
