% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XYomics-package.R
\docType{package}
\name{XYomics-package}
\alias{XYomics}
\alias{XYomics-package}
\title{XYomics: Analysis of Sex Differences in Omics Data}
\description{
The **XYomics** package provides functions for performing differential expression analysis,
pathway enrichment, gene regulatory network analysis, and comprehensive report generation for omics data.
}
\details{
This package is designed to integrate various omics analyses (e.g., functional omics and single-cell data)
with advanced visualization and reporting tools.
}
\author{
\strong{Maintainer}: Enrico Glaab \email{enrico.glaab@uni.lu}

Authors:
\itemize{
  \item Sophie Le Bars \email{sophie.lebars@uni.lu}
  \item Mohamed Soudy \email{mohamed.soudy@uni.lu}
}

Other contributors:
\itemize{
  \item Murodzhon Akhmedov [contributor, copyright holder]
}

}
\keyword{internal}
